% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{generateSample}
\alias{generateSample}
\title{Generates a sample from Tweedie Bayesian network}
\usage{
generateSample(p, phi, beta, n)
}
\arguments{
\item{p}{a vector of power parameters.}

\item{phi}{a vector of dispersion parameters.}

\item{beta}{a matrix of the Tweedie Bayesian network coefficients.}

\item{n}{an integer specifying the sample size.}
}
\value{
This function returns a matrix containing a random sample of n
vectors drawn form a Tweedie Bayesian network.
}
\description{
Generates an n-random sample according to a Tweedie Bayesian network with
parameters (p,phi,beta)
}
