% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{generateSample.tbn}
\alias{generateSample.tbn}
\title{Generates a sample from Tweedie Bayesian network}
\usage{
generateSample.tbn(beta, phi, p, n)
}
\arguments{
\item{beta}{a lower triangular matrix of the Tweedie Bayesian network coefficients.}

\item{phi}{a vector of dispersion parameters.}

\item{p}{a vector of power parameters.}

\item{n}{an integer specifying the sample size.}
}
\value{
This function returns a matrix containing a random sample of n
vectors drawn form a Tweedie Bayesian network.
}
\description{
Generates an n-random sample according to a Tweedie Bayesian network with
parameters (p,phi,beta)
}
\examples{
beta = matrix(0,ncol=5,nrow=5)
beta[1,1] = 1;
beta[2,1] = 0.5
beta[3,1] = -0.5; beta[3,2] = -0.7;beta[3,3] = -1.2
beta[4,1] = -2; beta[4,3] = -1
beta[5,1] = 0.2; beta[5,4] = 1; beta[5,5] = -2
generateSample.tbn(beta, c(1,2,1.5,1,1), c(2,2,3,2,0), 100)

}
