% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{learn.tbn}
\alias{learn.tbn}
\title{Learn a Tweedie Bayesian Network.}
\usage{
learn.tbn(data, pgrids, G = NULL, nodes.order = NULL, conf.level = 0.05)
}
\arguments{
\item{data}{a data frame containing the sample used to learn the model.}

\item{pgrids}{a list of numeric vectors containing the values among which
the power parameters are chosen in order to maximize the model's
log-likelihood. If the power parameter is knonw, each pgrid vector will
contain a single value.}

\item{G}{(optional) the Tweedie Bayesian network structure. If no graph is
provided, the structure will be learned relying on student statistical tests.}

\item{nodes.order}{(optional) The nodes ordering defining an ancestral
ordering of the Bayesian network.}

\item{conf.level}{(optioal) The confidence level for the structure learning.
Default value is 0.05.}
}
\value{
This function returns a named list consting of the Tweedie bayesian
network structure and parameters.
}
\description{
This function builds a Tweedie Bayesian network based on a data set.
It combines both structure and parameters learning. If a known graph is
provided, only the parameters will be estimated relying on the provided
network structure.
}
