% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{learn.trm}
\alias{learn.trm}
\title{Learn a Tweedie regression model parameters}
\usage{
learn.trm(data, formula, pgrid)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{pgrid}{a numeric vector containing the values among which the power
parameter p is chosen in order to maximize the model's log-likelihood.
If the power parameter is knonw, pgrid will contain a single value.}
}
\value{
This function returns a named list consisting of the Tweedie
regression parameters ("p" : the power parameter, "phi": the dispersion
parameter, "beta" : the regression coefficients and "glm": a regression
object as returned by the glm or lm functions.)
}
\description{
This is an implementation of the maximum likelihood method
to estimate the parameters of Tweedie regression model.
}
\examples{
learn.trm(data, V5~V2 + V3, pgrid=seq(2,3,0.05))

}
