% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{ll.tbn}
\alias{ll.tbn}
\title{Compute the log-likelihood within a Tweedie Bayesian network.}
\usage{
ll.tbn(beta, phi, p, X)
}
\arguments{
\item{beta}{a matrix of the Tweedie Bayesian network coefficients.}

\item{phi}{a vector of dispersion parameters.}

\item{p}{a vector of power parameters.}

\item{X}{a data frame containing the variables in the model.}
}
\value{
This function returns a named list consisting of the Tweedie
Bayesian network scores: Mean Squared Error (MSE), Root Mean Squared Error
(RMSE), Normalized Root Mean Squared Error (NRMSE) and the log-likelihood
(ll).
}
\description{
This function computes the log-likelihood within a Tweedie Bayesian network
for a given data set.
}
