% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{scores.tbn}
\alias{scores.tbn}
\title{Computes several likelihood-based score for a given Tweedie Bayesian Network.}
\usage{
scores.tbn(beta, phi, p, X)
}
\arguments{
\item{beta}{a matrix of the Tweedie Bayesian network coefficients.}

\item{phi}{a vector of dispersion parameters.}

\item{p}{a vector of power parameters.}

\item{X}{a data frame containing the variables in the model.}
}
\value{
This function returns a named list consisting of the Tweedie
Bayesian network scores: Akaike Information Criterion (AIC), Bayesian
Information Criterion (BIC) and the log-likelihood (ll).
}
\description{
Computes several likelihood-based score for a given Tweedie Bayesian Network:
the log-likelihood, the Akaike Information Criterion (AIC) and the Bayesian
Information Criterion (BIC).
}
