% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{scores.trm}
\alias{scores.trm}
\title{Compute likelihood-based score for a Tweedie regression model.}
\usage{
scores.trm(beta, phi, p, Y, X)
}
\arguments{
\item{beta}{a vector of regression coefficients.}

\item{phi}{the dispersion parameter of the Tweedie regression model.}

\item{p}{the power parameter of the Tweedie regression model.}

\item{Y}{a vector of responses.}

\item{X}{a matrix of covariates (including a vector of ones).}
}
\value{
This function returns a named list consisting of the Tweedie
regression model scores: Mean Squared Error (MSE), Root Mean Squared Error
(RMSE), Normalized Root Mean Squared Error (NRMSE) and the log-likelihood
(ll).
}
\description{
This function computes the log-likelihood of a Tweedie regression model given
a data set. It also provides the MSE, RMSE and NRMSE.
}
