% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{sensi.tbn}
\alias{sensi.tbn}
\title{Compute Sensitivity analyis given a tbn}
\usage{
sensi.tbn(beta, phi, p, db, X)
}
\arguments{
\item{beta}{a matrix of the Tweedie Bayesian network coefficients.}

\item{phi}{a vector of dispersion parameters.}

\item{p}{a vector of power parameters.}

\item{db}{a upper-triangular matrix of coefficients modification.}

\item{X}{a data frame containing the variables of the model.}
}
\value{
This function returns a double corresponding to the Tweedie Bayesian
network sensitivity.
}
\description{
Computes the sensitity of a Tweedie Bayesian network induced by a
modification of the coefficient matrix. This sensistivity relies on the
Kulback-Liebler divergence between the joint probability distributions
induced by the original and the modified Tweedie bayesian networks.
}
