% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{sensi.trm}
\alias{sensi.trm}
\title{Title}
\usage{
sensi.trm(beta, phi, p, db, X)
}
\arguments{
\item{beta}{a vector of regression coefficients.}

\item{phi}{the dispersion parameter of the Tweedie regression model.}

\item{p}{the power parameter of the Tweedie regression model.}

\item{db}{a vector of modification of the Tweedie regression coefficients.}

\item{X}{a data frame containing the variables in the model.}
}
\value{
This function returns the sensitivity of the tweedie regression model
induced by the regression coefficients modification.
}
\description{
Title
}
