% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{sensi.tbn}
\alias{sensi.tbn}
\title{Compute a tbn Sensitivity to coefficients modification}
\usage{
sensi.tbn(beta, phi, p, db, X)
}
\arguments{
\item{beta}{a lower triangular matrix of the Tweedie Bayesian network coefficients.}

\item{phi}{a vector of dispersion parameters.}

\item{p}{a vector of power parameters.}

\item{db}{a upper-triangular matrix of coefficients modification.}

\item{X}{a data frame containing the variables of the model.}
}
\value{
This function returns a double corresponding to the Tweedie Bayesian
network sensitivity.
}
\description{
Computes the sensitity of a Tweedie Bayesian network induced by a
modification of the coefficient matrix. This sensistivity relies on the
Kulback-Liebler divergence between the joint probability distributions
induced by the original and the modified Tweedie bayesian networks.
}
\examples{
\dontrun{
beta = matrix(0,ncol=5,nrow=5)
beta[1,1] = 1;
beta[2,1] = 0.5
beta[3,1] = -0.5; beta[3,2] = -0.7;beta[3,3] = -1.2
beta[4,1] = -2; beta[4,3] = -1
beta[5,1] = 0.2; beta[5,4] = 1; beta[5,5] = -2
db = matrix(0,ncol=5,nrow=5)
db[2,2] = 0.1 # adding an arc from X1 to X2.
sensi.tbn(beta, c(1,2,1.5,1,1), c(2,2,3,2,0), db, X)
}

}
