% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnfunctions.R
\name{sensi.trm}
\alias{sensi.trm}
\title{Compute the sensitivity of TRM to coefficients modification.}
\usage{
sensi.trm(beta, phi, p, db, X)
}
\arguments{
\item{beta}{a vector of regression coefficients.}

\item{phi}{the dispersion parameter of the Tweedie regression model.}

\item{p}{the power parameter of the Tweedie regression model.}

\item{db}{a vector of modification of the Tweedie regression coefficients.}

\item{X}{a data frame containing the variables in the model.}
}
\value{
This function returns the sensitivity of the tweedie regression model
induced by the regression coefficients modification.
}
\description{
Compute the sensitivity of TRM to coefficients modification. The returned
value corresponds to the local sensitivity of a Tweedie Bayesian network.
}
\examples{
\dontrun{
sensi.trm(c(0.2,0,0,1,-2),1,2, c(-0.1,0,0,-0.2,0), X)
}


}
